+++ scheduler/models/__init__.py
@@ -1,5 +1,5 @@
 from .args import TaskKwarg, TaskArg
-from .scheduled_task import BaseTask, ScheduledTask, RepeatableTask, CronTask
+from .old_scheduled_task import BaseTask, ScheduledTask, RepeatableTask, CronTask
 from .queue import Queue
 from .task import Task
